[AOT Refactor] alias runtime wrappers
[ghstack-poisoned]
