+++ torch/_functorch/aot_autograd.py
@@ -1,6 +1,5 @@
 import itertools
 import logging
-import pprint
 from contextlib import nullcontext
 from functools import partial, wraps
 from typing import Any, Callable, Dict, List, Optional, Tuple
@@ -28,7 +27,7 @@ from torch.utils._python_dispatch import is_traceable_wrapper_subclass
 from torch._decomp.decompositions_for_rng import PhiloxStateTracker, rng_decompositions
 from . import config
 from .partitioners import default_partition
-from torch._guards import TracingContext, DuplicateInputs
+from torch._guards import TracingContext
 
 from ._aot_autograd.utils import (  # noqa: F401
     strict_zip, _get_symint_hints, create_tree_flattened_fn,
@@ -69,16 +68,9 @@ from ._aot_autograd.collect_metadata_analysis import (
     run_functionalized_fw_and_collect_metadata,
 )
 from ._aot_autograd.input_output_analysis import (
-    remove_dupe_metadata,
-    merge_view_inputs,
-    create_synthetic_base_metadata,
     create_graph_signature,
 )
-from ._aot_autograd.runtime_wrappers import (
-    create_runtime_wrapper,
-    functionalized_rng_runtime_epilogue,
-    aot_dispatch_subclass_wrapper,
-)
+
 from ._aot_autograd.traced_function_transforms import (
     fn_input_mutations_to_outputs,
     fn_prepped_for_autograd,
@@ -86,6 +78,15 @@ from ._aot_autograd.traced_function_transforms import (
     create_functionalized_fn,
     aot_dispatch_subclass,
     create_functional_call,
+)
+from ._aot_autograd.runtime_wrappers import (
+    create_runtime_wrapper,
+    functionalized_rng_runtime_epilogue,
+    aot_dispatch_subclass_wrapper,
+)
+from ._aot_autograd.alias_runtime_wrappers import (
+    aot_wrapper_dedupe,
+    aot_wrapper_synthetic_base,
 )
 
 zip = strict_zip
@@ -474,412 +475,6 @@ def aot_dispatch_base(flat_fn, flat_args: List[Tensor], aot_config: AOTConfig, *
     )
 
     return compiled_fn
-
-
-# MOTIVATION:
-#
-# When tracing functions for future execution, one must be careful not to pass
-# in the same input tensor multiple times (e.g., f(x, x), as this can result
-# in graphs that are ONLY valid if you later pass a new tensor in exactly the
-# same way (e.g., f(y, y)).  (NB: we really mean duplicate; two distinct
-# tensors that alias each other is a different situation that is covered by
-# aot_dispatch_deduplicated_autograd). Here are two examples:
-#
-# (1) Suppose you have a function:
-#
-#   def f(x, y):
-#       return x + y
-#
-# If you make_fx(f)(x, x), you will trace out:
-#
-#   def f(x, y):
-#       return y + y
-#
-# Oops!
-#
-# (2) For most tensors x and y, you can compute f's gradient with respect to
-# these to inputs by saying torch.autograd.grad(f(x, y), (x, y)).  However,
-# if x is y, you will trace out a program that gets incorrect gradients:
-#
-#   >>> x = torch.randn(1, requires_grad=True)
-#   >>> torch.autograd.grad(x + x, (x, x))
-#   (tensor([2.]), tensor([2.]))
-#
-# In other words, the gradient is double-counted.  Deduplicating the arguments
-# gives you an appropriate gradient:
-#
-#   >>> y = torch.randn(1, requires_grad=True)
-#   >>> torch.autograd.grad(x + y, (x, y))
-#   (tensor([1.]), tensor([1.]))
-#
-# HOW TO DEDUPLICATE:
-#
-# There are a few strategies, in order of preference:
-#
-# 1. For every duplicate argument to the function, detach it into
-#    a separate leaf tensor, so that it is no longer duplicated.
-#
-#       PRO: The resulting compiled graph works for any configuration
-#       of duplicated arguments.
-#
-#       CON: It does not (naively) work if you mutate the metadata of inputs:
-#
-#           def f(x, y):
-#               x.transpose_(0, 1)
-#               y.transpose_(0, 2)
-#
-#           x = torch.randn(2, 3, 4)
-#           f(x, x)
-#
-#       The ordering of the transposes inside f dictates whether or not
-#       you get [4, 2, 3] or [3, 4, 2].  This means that you cannot precompute
-#       what metadata mutations should get applied to each input; you need to
-#       assume they aren't duplicates (what we do today) or preserve
-#       the original metadata mutations exactly in order, so that they work
-#       for any duplicate configuration.
-#
-#       CON: It does not (naively) work if you mutate the data of inputs.
-#       In particular, leaf tensors that require grad cannot be mutated,
-#       this makes it impossible to differentiate with respect to the original
-#       base.
-#
-# 2. For every duplicate argument to the function, remove it, so it is
-#    no longer part of the "true" signature:
-#
-#       PRO: Implemented naively, it still works for metadata/data mutation.
-#
-#       CON: The resulting compiled graph is duplicate-specialized: it only
-#       works if future calls duplicate arguments in exactly the same way.
-#       Horribly, Dynamo doesn't guard on this at the moment.  But even if
-#       it did, you could still end up recompiling a bunch of each duplicate.
-#
-# Our strategy is to do (1) if we can, and do (2) otherwise, erroring if
-# Dynamo's guards are not enough.  In practice, this seems to cover
-# everything.
-#
-def aot_wrapper_dedupe(
-    flat_fn,
-    flat_args: List[Tensor],
-    aot_config: AOTConfig,
-    *,
-    compiler_fn,
-    fw_metadata,
-):
-    # Use information about whether or not flat_fn mutates its arguments
-    # or not to handle dupe args
-
-    # Strategy 1: For any input that is not mutated, we can leafify it if we
-    # need to remove a duplicate.
-    leaf_flat_args = []
-    args_set = set()
-    ok = True
-
-    for i, a in enumerate(flat_args):
-        if not isinstance(a, torch.Tensor):
-            leaf_flat_args.append(a)
-        elif a not in args_set:
-            args_set.add(a)
-            leaf_flat_args.append(a)
-        elif not fw_metadata.input_info[i].mutates_data and not fw_metadata.input_info[i].mutates_metadata:
-            leaf_flat_args.append(a.detach().requires_grad_(a.requires_grad))
-        else:
-            ok = False
-            break
-
-    if ok:
-        return compiler_fn(flat_fn, leaf_flat_args, aot_config, fw_metadata=fw_metadata)
-
-    if requires_subclass_dispatch(leaf_flat_args, fw_metadata):
-        raise RuntimeError("""\
-Encountered duplicate inputs that are mutated in the graph, but at least one input/output
-to the graph is a tensor subclass. This is not supported today. You can try to
-remove the aliasing yourself as a workaround, or otherwise file an issue on github.""")
-
-    # export path: ban duplicate inputs for now, add later if requested.
-    if aot_config.is_export:
-        raise RuntimeError(f"""\
-Encountered duplicated inputs that are mutated in the graph you are trying to export.
-This functionality is currently not supported. If needed, please file a github issue.
-
-fw_metadata={str(fw_metadata)}
-        """)
-
-    # Strategy 2: Duplicate specialize.
-    #
-    # In Haskell types, suppose you have:
-    #
-    #   add_dupe_args :: DedupedArgs -> Args
-    #   remove_dupe_args :: Args -> DedupedArgs
-    #
-    #   compiler_fn
-    #       :: (DedupedArgs -> R) -> DedupedArgs -> AOTConfig -> (DedupedArgs -> R)
-    #   deped_compiler_fn
-    #       :: (Args -> R) -> Args -> AOTConfig -> (Args -> R)
-    #
-    # Then the code below can be written in point-free style as:
-    #
-    #   deduped_compiler_fn f a c =
-    #       compiler_fn (f . add_dupe_args) (remove_dupe_args a) c . remove_dupe_args
-    #
-    # Suppose you have:
-    #
-    #   [a, b, a, c]
-    #
-    # We want:
-    #
-    #   remove_dupe_args([a, b, a, c]) == [a, b, c]
-    #   add_dupe_args([a, b, c]) == [a, b, a, c]
-    #
-    # This is done via (respectively):
-    #
-    #   seen_args = {a: 0, b: 1, c: 2}
-    #   enumerate(add_dupe_map) = [  # how to get args from the deduped list
-    #       (0, 0),
-    #       (1, 1),
-    #       (2, 0),
-    #       (3, 2),
-    #   ]
-    #   keep_arg_mask = [True, True, False, True]
-
-    seen_args = {}
-    keep_arg_mask = []
-    # Implicitly map duped arg position (list index) to de-duped arg position
-    add_dupe_map: List[int] = []
-    duped_arg_len = len(flat_args)
-
-    j = 0  # index into deduped_flat_args
-    for t in flat_args:
-        if isinstance(t, torch.Tensor):
-            if t in seen_args:
-                keep_arg_mask.append(False)
-                add_dupe_map.append(seen_args[t])
-                continue
-            seen_args[t] = j
-
-        keep_arg_mask.append(True)
-        add_dupe_map.append(j)
-        j += 1
-    assert len(add_dupe_map) == duped_arg_len, (
-        f"Expects add_dupe_map to have length {duped_arg_len} but got {len(add_dupe_map)}"
-    )
-
-    # NB: Hot path, avoid set lookups here
-    # TODO: Can avoid the zip here too, probably
-    def remove_dupe_args(args):
-        return [t for t, keep in zip(args, keep_arg_mask) if keep]
-
-    def add_dupe_args(args):
-        return [args[add_dupe_map[i]] for i in range(duped_arg_len)]
-
-    deduped_flat_args = remove_dupe_args(flat_args)
-
-    # Update our input metadata to remove duped input metadata.
-    updated_fw_metadata = remove_dupe_metadata(fw_metadata, keep_arg_mask, add_dupe_map)
-
-    if tracing_context := TracingContext.try_get() and aot_config.aot_autograd_arg_pos_to_source:
-        # TODO(voz): This structure is 1:1, we could consider an alternate structure like
-        # kept_pos:[dupe_arg_pos], however, add_dupe_map is 1:1 so we would need a new structure there,
-        # which feels like needless complexity for a tiny bit of efficiency at this point.
-        for dupe_arg_pos, (kept_pos, keep_arg) in enumerate(zip(add_dupe_map, keep_arg_mask)):
-            if not keep_arg:
-                dupe_arg_source = aot_config.aot_autograd_arg_pos_to_source[dupe_arg_pos]
-                kept_arg_source = aot_config.aot_autograd_arg_pos_to_source[kept_pos]
-                tracing_context.guards_context.aotautograd_guards.append(DuplicateInputs(kept_arg_source, dupe_arg_source))
-
-    @wraps(flat_fn)
-    def wrapped_flat_fn(*args):
-        return flat_fn(*add_dupe_args(args))
-
-    if config.debug_assert:
-        ref_fw_metadata = run_functionalized_fw_and_collect_metadata(
-            wrapped_flat_fn,
-            keep_input_mutations=fw_metadata.keep_input_mutations,
-            is_train=fw_metadata.is_train,
-        )(*deduped_flat_args)
-        assert ref_fw_metadata == updated_fw_metadata, \
-            f'ref_metadata={str(ref_fw_metadata)}, actual_metadata={str(updated_fw_metadata)}'
-
-    compiled_fn = compiler_fn(wrapped_flat_fn, deduped_flat_args, aot_config, fw_metadata=updated_fw_metadata)
-
-    if not hasattr(compiled_fn, "_boxed_call"):
-        compiled_fn = make_boxed_func(compiled_fn)
-
-    @wraps(compiled_fn)
-    def wrapped_compiled_fn(args):
-        deduped_args = remove_dupe_args(args)
-        args.clear()
-        return compiled_fn(deduped_args)
-
-    wrapped_compiled_fn._boxed_call = True
-
-    # This can be uncommented when we properly guard for duplicates,
-    # but right now we must not do it.
-    # if not config.debug_assert:
-    #     return wrapped_compiled_fn
-
-    @wraps(wrapped_compiled_fn)
-    def debugged_compiled_fn(args):
-        # Test that the computed remove/add arg functions are an inverse
-        new_args = add_dupe_args(remove_dupe_args(args))
-        seen = {}
-        for i, (x, y) in enumerate(zip(new_args, args)):
-            seen[y] = None
-            assert x is y, format_guard_bug_msg(
-                aot_config,
-                f"{describe_input(i, aot_config)} would be a duplicate of "
-                f"{describe_input(add_dupe_map[i], aot_config)}",
-            )
-        # This is only an error if there is metadata mutation on both of
-        # the duped arguments; in this case, we need to know what order
-        # the metadata mutation applies in.  You'll get the correct result
-        # otherwise, because a graph that assumes distinct inputs works if
-        # you dupe the inputs (the gradient contributions from each input
-        # will get summed up appropriately.)
-        #
-        # TODO: work out how to setup this assert correctly
-        """
-        assert len(seen) == unique_args, format_guard_bug_msg(aot_config,
-            f"there would be {unique_args} distinct arguments"
-        )
-        """
-        return wrapped_compiled_fn(args)
-
-    debugged_compiled_fn._boxed_call = True
-
-    return debugged_compiled_fn
-
-# This layer handles the situation where you have two inputs that alias each other,
-# and one of the inputs is mutated.
-# We need to take special care to ensure that the mutation is applied to the other aliases in the graph.
-#
-# pre-condition: aot_wrapper_dedup has already run.
-# (This function will in theory work if there are duplicate args.
-# However, the synthetic base code path is a bit sub-optimal, and running with dupe'd inputs
-# would cause us to hit that path more frequently).
-def aot_wrapper_synthetic_base(
-    flat_fn,
-    flat_args: List[Tensor],
-    aot_config: AOTConfig,
-    *,
-    fw_metadata: ViewAndMutationMeta,
-    # Currently, the only reason we need to plumb this bool is because
-    # the synthetic base code prohibits more cases in the autograd case than the inference case.
-    needs_autograd: bool,
-    compiler_fn,
-):
-    is_inference = not needs_autograd
-    flat_args_with_synthetic_bases, synthetic_base_info = merge_view_inputs(
-        flat_args, fw_metadata.input_info, is_inference=is_inference,
-    )
-    # Happy path: we don't need synthetic bases
-    if synthetic_base_info is None:
-        return compiler_fn(flat_fn, flat_args, aot_config, fw_metadata=fw_metadata)
-
-    # export path: ban synthetic bases for now, add later if requested.
-    if requires_subclass_dispatch(flat_args, fw_metadata):
-        raise RuntimeError("""\
-Encountered aliased inputs that are mutated in the graph, but at least one input/output
-to the graph is a tensor subclass. This is not supported today. You can try to
-remove the aliasing yourself as a workaround, or otherwise file an issue on github.""")
-
-    if aot_config.is_export:
-        raise RuntimeError(f"""\
-Encountered aliased inputs that are mutated in the graph you are trying to export.
-This functionality is currently not supported. If needed, please file a github issue.
-
-synthetic_base_info={str(synthetic_base_info)}
-
-fw_metadata={str(fw_metadata)}
-        """)
-
-    assert len(fw_metadata.input_info) == len(synthetic_base_info)
-
-    # Update our forward metadata to take synthetic bases into account
-    fw_metadata_updated, aliased_arg_idx_with_metadata_mutations = \
-        create_synthetic_base_metadata(fw_metadata, synthetic_base_info, flat_args, flat_args_with_synthetic_bases)
-
-    num_aliased_args_with_metadata_mutations = len(aliased_arg_idx_with_metadata_mutations)
-
-    def unpack_synthetic_bases(primals: List[Any]) -> List[Any]:
-        f_args_inner = []
-        for inner_idx_or_tuple in synthetic_base_info:
-            if isinstance(inner_idx_or_tuple, int):
-                f_args_inner.append(primals[inner_idx_or_tuple])
-            else:
-                inner_base_idx, view_tensor = inner_idx_or_tuple
-                base = primals[inner_base_idx]
-                view_arg = gen_alias_from_base(
-                    base, view_tensor, view_tensor.requires_grad
-                )
-                f_args_inner.append(view_arg)
-        return f_args_inner
-
-    @wraps(flat_fn)
-    def wrapped_flat_fn(*args):
-        unpacked_args = unpack_synthetic_bases(args)
-        # This is a bit subtle. The goal of this entire function (aot_dispatch_synthetic_bases)
-        # is to relieve the downstream logic from having to reason about mutations on inputs that alias
-        # each other, by replacing aliased inputs with a synthetic base.
-        # One area where this breaks down a bit however is if one of those aliased inputs
-        # experienced a metadata mutation.
-        # We are now obligated to reapply the metadata mutation directly to the user's input;
-        # it isn't enough to apply mutations back to the synthetic base in the downstream logic.
-        #
-        # The way we handle this is by pretending that those aliased inputs that experience metadata mutations
-        # are additional outputs in the user's forward function.
-        # The downstream logic will just treat these as "user outputs that alias inputs".
-        # However, we will manually grab them at runtime here, use them to reapply the metadata mutation
-        # to the user inputs, and not return them to the user.
-        aliased_args_with_metadata_mutations = [
-            x for i, x in enumerate(unpacked_args) if i in aliased_arg_idx_with_metadata_mutations]
-        if len(aliased_args_with_metadata_mutations) > 0:
-            return *(flat_fn(*unpacked_args)), *aliased_args_with_metadata_mutations
-        else:
-            return flat_fn(*unpacked_args)
-
-    if config.debug_assert:
-        ref_fw_metadata = run_functionalized_fw_and_collect_metadata(
-            wrapped_flat_fn,
-            keep_input_mutations=fw_metadata.keep_input_mutations,
-            is_train=fw_metadata.is_train,
-        )(*flat_args_with_synthetic_bases)
-        assert ref_fw_metadata == fw_metadata_updated, (
-            f'ref_metadata={pprint.pformat(partial_flatten_asdict(ref_fw_metadata))}, '
-            f'\nactual_metadata={pprint.pformat(partial_flatten_asdict(fw_metadata_updated))}'
-        )
-
-    compiled_fn = compiler_fn(wrapped_flat_fn, flat_args_with_synthetic_bases, aot_config, fw_metadata=fw_metadata_updated)
-
-    if not hasattr(compiled_fn, "_boxed_call"):
-        compiled_fn = make_boxed_func(compiled_fn)
-
-    @wraps(compiled_fn)
-    def wrapped_compiled_fn(args):
-        args_with_synthetic_bases, synthetic_base_info = merge_view_inputs(
-            args, fw_metadata.input_info, is_inference=is_inference
-        )
-        assert synthetic_base_info is not None
-        aliased_args_w_metadata_mutations = [args[i] for i in aliased_arg_idx_with_metadata_mutations]
-        args.clear()
-        outs = compiled_fn(args_with_synthetic_bases)
-        if num_aliased_args_with_metadata_mutations > 0:
-            # This code does not handle **all** input metadata mutations.
-            # Instead, it only handles metadata mutations on inputs that were converted into synthetic bases
-            # (which only happens if at least one aliased input experienced a data mutation).
-            # e.g:
-            # def f(a, b):
-            #     a.mul_(2)
-            #     b.t_(1, 0)
-            # f(x.view(2, 2), x.view(2, 2))
-            mutated_metadata_inps = outs[-num_aliased_args_with_metadata_mutations:]
-            user_outs = outs[:-num_aliased_args_with_metadata_mutations]
-            for inp, mutated_inp in zip(aliased_args_w_metadata_mutations, mutated_metadata_inps):
-                inp.as_strided_(mutated_inp.size(), mutated_inp.stride(), mutated_inp.storage_offset())
-            return user_outs
-        return outs
-
-    return wrapped_compiled_fn
-
 
 # Has the precondition that there
 # are no duplicate arguments in flat_args (e.g., the same Tensor
